+++ NEWS.md
@@ -5,4 +5,5 @@ version 0.1.0.9000
 
 * Massive rewrite with help of [Appsilon Data Science](https://appsilondatascience.com/) developers, @DamianRodziewicz and @mateuszwyszynski.
 * Key changes are package is primarily designed around building new apps on the fly, via `newApp`.
+* new module `authGoogle` allows logging in and authenticating users against a whitelist.
 * Much of the old code is obsolete and is largely deprecated/defunct.